refactor: 랜딩페이지 관련 컴포넌트 home 디렉토리로 이동
