+++ components/home/Slider.jsx
